Instrumentation support with OnFlush and OnDrop
OnDrop is called if one or more entries are dropped
and are lost. Shouldn't happen if MayDrop flag is false.

OnFlush is called after the entries are sent (flushed)
to the LogDNA, with an error that indicates whenever
that was a success or not. Note, the failure here does
not indicate that the entries were lost.